+++ examples/numpy_file_read.cpp
@@ -1,5 +1,5 @@
 // Your First C++ Program
-#include "aare/FileHandler.hpp"
+#include "aare/ContextManager.hpp"
 #include <iostream>
 
 #define AARE_ROOT_DIR_VAR "PROJECT_ROOT_DIR"
